# Fix compositing performance (3x improvement)