--- a/src/main/scripts/neo4jd3.js
+++ b/src/main/scripts/neo4jd3.js
@@ -646,57 +646,6 @@ function Neo4jD3(_selector, _options) {
         return graph;
     }
 
-    function randomD3Data(d, maxNodesToGenerate) {
-        var data = {
-                nodes: [],
-                relationships: []
-            },
-            i,
-            label,
-            node,
-            numNodes = (maxNodesToGenerate * Math.random() << 0) + 1,
-            relationship,
-            s = size();
-
-        for (i = 0; i < numNodes; i++) {
-            label = randomLabel();
-
-            node = {
-                id: s.nodes + 1 + i,
-                labels: [label],
-                properties: {
-                    random: label
-                },
-                x: d.x,
-                y: d.y
-            };
-
-            data.nodes[data.nodes.length] = node;
-
-            relationship = {
-                id: s.relationships + 1 + i,
-                type: label.toUpperCase(),
-                startNode: d.id,
-                endNode: s.nodes + 1 + i,
-                properties: {
-                    from: Date.now()
-                },
-                source: d.id,
-                target: s.nodes + 1 + i,
-                linknum: s.relationships + 1 + i
-            };
-
-            data.relationships[data.relationships.length] = relationship;
-        }
-
-        return data;
-    }
-
-    function randomLabel() {
-        var icons = Object.keys(options.iconMap);
-        return icons[icons.length * Math.random() << 0];
-    }
-
     function rotate(cx, cy, x, y, angle) {
         var radians = (Math.PI / 180) * angle,
             cos = Math.cos(radians),
@@ -1033,7 +982,6 @@ function Neo4jD3(_selector, _options) {
 
     return {
         neo4jDataToD3Data: neo4jDataToD3Data,
-        randomD3Data: randomD3Data,
         size: size,
         updateWithD3Data: updateWithD3Data,
         updateWithNeo4jData: updateWithNeo4jData,