### How to test
Comment out `tickRelationshipOutlines();` in `neo4jd3.js` like this:
``` js
            tickRelationshipsTexts();
            //tickRelationshipsOutlines();
            tickRelationshipsOverlays();
```

This removes other bottlenecks. Compare the performance of dragging around icons before vs after this change.

Before (75ms total task time~, around 50ms is compositing):
![image](https://user-images.githubusercontent.com/7528322/165595115-f3bb3598-278d-42b4-ad0b-09ec5c079740.png)

After (25ms total task time~, around 1ms is composting):
![image](https://user-images.githubusercontent.com/7528322/165595521-bb3fbdb2-4406-48c0-93d3-81164a8a370e.png)

Reference: https://stackoverflow.com/a/40533907/3991315