@BurkusCat 2022-04-27
Improves #8